pyproject.toml
[project]
name = "greekapp"
version = "0.1.0"
description = "Greek vocabulary learning app with spaced repetition and natural daily interactions"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.4"]
deploy = [
    "anthropic>=0.40",
    "httpx>=0.27",
    "flask>=3.0",
    "gunicorn>=22.0",
    "psycopg2-binary>=2.9",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
include = ["greekapp*"]

[project.scripts]
greek = "greekapp.cli:cli"
